boosted CT log checking and fixed bug with CT entry id which are Int64, not Int32

+++ SSLVerifier.Core/CertTransparency/CertShLogEntry.cs
@@ -4,9 +4,9 @@ using Newtonsoft.Json;
 namespace SSLVerifier.Core.CertTransparency {
     class CertShLogEntry : ICertLogEntry {
         [JsonProperty("id")]
-        public Int32 EntryID { get; set; }
+        public Int64 EntryID { get; set; }
         [JsonProperty("issuer_ca_id")]
-        public Int32 IssuerID { get; set; }
+        public Int64 IssuerID { get; set; }
         [JsonProperty("issuer_name")]
         public String IssuerName { get; set; }
         [JsonProperty("name_value")]
