+++ SSLVerifier.Core/CertTransparency/ICertLogEntry.cs
@@ -2,8 +2,8 @@
 
 namespace SSLVerifier.Core.CertTransparency {
     public interface ICertLogEntry {
-        Int32 EntryID { get; }
-        Int32 IssuerID { get; }
+        Int64 EntryID { get; }
+        Int64 IssuerID { get; }
         String IssuerName { get; }
         String HostName { get; }
         DateTime Timestamp { get; }
